Maintenance #CHANGE: Split code files containing tons of enums, so each enum is no located in a separate code file.
