+++ Duality/Drawing/Enums/BlendMode.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+using Duality.Editor;
+using Duality.Resources;
+
+namespace Duality.Drawing
+{
+	/// <summary>
+	/// Enumerates different behviours on how to blend color data onto existing background color.
+	/// </summary>
+	/// <seealso cref="Duality.Resources.DrawTechnique"/>
+	public enum BlendMode
+	{
+		/// <summary>
+		/// When passing this to a method, this value can be used to indicate "Restore to default settings".
+		/// </summary>
+		[EditorHintFlags(MemberFlags.Invisible)]
+		Reset = -1,
+
+		/// <summary>
+		/// Incoming color overwrites background color completely. Doesn't need Z-Sorting.
+		/// </summary>
+		Solid,
+		/// <summary>
+		/// Incoming color overwrites background color but leaves out areas with low alpha. Doesn't need Z-Sorting.
+		/// </summary>
+		Mask,
+
+		/// <summary>
+		/// Incoming color is multiplied by its alpha value and then added to background color. Needs Z-Sorting.
+		/// </summary>
+		Add,
+		/// <summary>
+		/// Incoming color overwrites background color, weighted by its alpha value. Needs Z-Sorting.
+		/// </summary>
+		Alpha,
+		/// <summary>
+		/// Premultiplied Alpha: Colors specify brightness, alpha specifies opacity. Needs Z-Sorting.
+		/// </summary>
+		AlphaPre,
+		/// <summary>
+		/// Incoming color scales background color. Needs Z-Sorting.
+		/// </summary>
+		Multiply,
+		/// <summary>
+		/// Incoming color is multiplied and then added to background color. Needs Z-Sorting.
+		/// </summary>
+		Light,
+		/// <summary>
+		/// Incoming color inverts background color. Needs Z-Sorting.
+		/// </summary>
+		Invert,
+
+		/// <summary>
+		/// The total number of available BlendModes.
+		/// </summary>
+		[EditorHintFlags(MemberFlags.Invisible)]
+		Count
+	}
+}
