+++ Duality/Drawing/Enums/DashPattern.cs
