+++ Duality/Drawing/Enums/RenderMatrix.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Duality.Drawing
+{
+	/// <summary>
+	/// Specifies a rendering matrix setup.
+	/// </summary>
+	public enum RenderMatrix
+	{
+		PerspectiveWorld,
+		OrthoScreen
+	}
+}
