+++ Duality/Drawing/Enums/ShaderType.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Duality.Drawing
+{
+	public enum ShaderType
+	{
+		Vertex,
+		Fragment
+	}
+}
