+++ Duality/Duality.csproj
@@ -85,12 +85,27 @@
     <Compile Include="Cloning\ICloneTargetSetup.cs" />
     <Compile Include="Cloning\CloneType.cs" />
     <Compile Include="Cloning\ManuallyClonedAttribute.cs" />
+    <Compile Include="Components\ICmpInterfaces\ICmpEditorUpdatable.cs" />
+    <Compile Include="Components\ICmpInterfaces\ICmpInitializable.cs" />
+    <Compile Include="Components\ICmpInterfaces\ICmpCollisionListener.cs" />
+    <Compile Include="Components\ICmpInterfaces\ICmpRenderer.cs" />
+    <Compile Include="Components\ICmpInterfaces\ICmpUpdatable.cs" />
     <Compile Include="Components\Physics\CollisionCategory.cs" />
+    <Compile Include="Drawing\Enums\BlendMode.cs" />
+    <Compile Include="Drawing\Enums\ClearFlag.cs" />
+    <Compile Include="Drawing\Enums\ShaderType.cs" />
+    <Compile Include="Drawing\Enums\RenderMatrix.cs" />
+    <Compile Include="Drawing\Enums\PerspectiveMode.cs" />
+    <Compile Include="Drawing\Enums\VertexMode.cs" />
     <Compile Include="Drawing\IDrawBatch.cs" />
+    <Compile Include="Drawing\Enums\TextureSizeMode.cs" />
+    <Compile Include="Drawing\Enums\TextureMagFilter.cs" />
+    <Compile Include="Drawing\Enums\TextureMinFilter.cs" />
+    <Compile Include="Drawing\Enums\TextureWrapMode.cs" />
     <Compile Include="Drawing\VertexFormatDefinition.cs" />
-    <Compile Include="Drawing\TextureEnums.cs" />
-    <Compile Include="Drawing\IRendererVisibilityStrategy.cs" />
-    <Compile Include="Drawing\DefaultRendererVisibilityStrategy.cs" />
+    <Compile Include="Drawing\Enums\TexturePixelFormat.cs" />
+    <Compile Include="Components\Other\IRendererVisibilityStrategy.cs" />
+    <Compile Include="Components\Other\DefaultRendererVisibilityStrategy.cs" />
     <Compile Include="Drawing\ScreenResolution.cs" />
     <Compile Include="Input\UserInputEventArgs.cs" />
     <Compile Include="Input\JoystickButtonEventArgs.cs" />
@@ -117,7 +132,7 @@
     <Compile Include="Cloning\Surrogates\DelegateSurrogate.cs" />
     <Compile Include="Cloning\Surrogates\DictionarySurrogate.cs" />
     <Compile Include="Components\Diagnostics\VisualLogRenderer.cs" />
-    <Compile Include="Components\ICmpInterfaces.cs" />
+    <Compile Include="Components\Other\CollisionData.cs" />
     <Compile Include="Components\Physics\Joints\GearJointInfo.cs" />
     <Compile Include="Components\Physics\Joints\PulleyJointInfo.cs" />
     <Compile Include="Components\Physics\Joints\LineJointInfo.cs" />
@@ -147,9 +162,9 @@
     <Compile Include="Properties\CoreResNames.cs" />
     <Compile Include="Drawing\DrawDevice.cs" />
     <Compile Include="Drawing\IDrawDevice.cs" />
-    <Compile Include="Drawing\DrawDeviceEnums.cs" />
+    <Compile Include="Drawing\Enums\VisibilityFlag.cs" />
     <Compile Include="Drawing\CanvasBuffer.cs" />
-    <Compile Include="Drawing\DashPattern.cs" />
+    <Compile Include="Drawing\Enums\DashPattern.cs" />
     <Compile Include="Drawing\CanvasState.cs" />
     <Compile Include="Utility\Math\Alignment.cs" />
     <Compile Include="Components\Renderers\AnimSpriteRenderer.cs" />
